+++ bigdata-test/src/main/java/io/tidb/bigdata/test/StreamIntegrationTest.java
@@ -1,0 +1,5 @@
+package io.tidb.bigdata.test;
+
+public interface StreamIntegrationTest {
+
+}
