+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/delete/TiKVDeleteNotSupportTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.tidb.bigdata.flink.tidb.stream.delete;
+package io.tidb.bigdata.flink.tidb.delete;
 
 import static io.tidb.bigdata.flink.connector.TiDBOptions.DELETE_ENABLE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.IGNORE_PARSE_ERRORS;
@@ -33,6 +33,7 @@ import io.tidb.bigdata.flink.connector.TiDBOptions.SinkTransaction;
 import io.tidb.bigdata.flink.tidb.FlinkTestBase;
 import io.tidb.bigdata.test.IntegrationTest;
 import io.tidb.bigdata.test.RandomUtils;
+import io.tidb.bigdata.test.StreamIntegrationTest;
 import io.tidb.bigdata.tidb.TiDBWriteMode;
 import java.util.Arrays;
 import java.util.Collection;
@@ -47,7 +48,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized.Parameters;
 
 /** append mode will throw exception Global is not tested for it can't work with streaming */
-@Category(IntegrationTest.class)
+@Category(StreamIntegrationTest.class)
 @RunWith(org.junit.runners.Parameterized.class)
 public class TiKVDeleteNotSupportTest extends FlinkTestBase {
