+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/delete/TiKVDeleteTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package io.tidb.bigdata.flink.tidb.stream;
+package io.tidb.bigdata.flink.tidb.delete;
 
-import static io.tidb.bigdata.flink.connector.TiDBOptions.DEDUPLICATE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.DELETE_ENABLE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.IGNORE_PARSE_ERRORS;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_BUFFER_SIZE;
@@ -34,7 +33,10 @@ import io.tidb.bigdata.flink.connector.TiDBOptions.SinkTransaction;
 import io.tidb.bigdata.flink.tidb.FlinkTestBase;
 import io.tidb.bigdata.test.IntegrationTest;
 import io.tidb.bigdata.test.RandomUtils;
+import io.tidb.bigdata.test.StreamIntegrationTest;
 import io.tidb.bigdata.tidb.TiDBWriteMode;
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ExecutorService;
@@ -49,13 +51,43 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized.Parameters;
 
-@Category(IntegrationTest.class)
-public class TiKVDeduplicateTest extends FlinkTestBase {
+/** Test for pk, uk and multiple_uk table */
+@Category(StreamIntegrationTest.class)
+@RunWith(org.junit.runners.Parameterized.class)
+public class TiKVDeleteTest extends FlinkTestBase {
 
   private String srcTable;
 
   private String dstTable;
+
+  public TiKVDeleteTest(String flinkDeleteTable, int result, String kafkaGroup) {
+    this.flinkDeleteTable = flinkDeleteTable;
+    this.result = result;
+    this.kafkaGroup = kafkaGroup;
+  }
+
+  @Parameters(name = "{index}: FlinkDeleteTable={0}, Result={1} ,KafkaGroup={2} ")
+  public static Collection<Object[]> data() {
+    return Arrays.asList(
+        new Object[][] {
+          {TABLE_PK, 2, "group_d_1"}, {TABLE_UK, 2, "group_d_2"}, {TABLE_MUTILE_UK, 2, "group_d_3"},
+        });
+  }
+
+  private final String flinkDeleteTable;
+  private final int result;
+  private final String kafkaGroup;
+
+  private static final String TABLE_PK =
+      "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
+          + "(\n"
+          + "    c1  varchar(255),\n"
+          + "    c2  varchar(255),\n"
+          + "    PRIMARY KEY (`c1`) \n"
+          + ")";
 
   private static final String TABLE_UK =
       "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
@@ -66,10 +98,18 @@ public class TiKVDeduplicateTest extends FlinkTestBase {
           + "    UNIQUE INDEX `uniq_2`(`c2`) USING BTREE\n"
           + ")";
 
+  private static final String TABLE_MUTILE_UK =
+      "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
+          + "(\n"
+          + "    c1  bigint(20) NOT NULL,\n"
+          + "    c2  bigint(20) NOT NULL,\n"
+          + "    UNIQUE INDEX `uniq_1`(`c1`,`c2`) USING BTREE\n"
+          + ")";
+
   @Test
-  public void testDeduplicate() throws Exception {
-    srcTable = "flink_deduplicate_src_test" + RandomUtils.randomString();
-    dstTable = "flink_deduplicate_dst_test" + RandomUtils.randomString();
+  public void testDelete() throws Exception {
+    srcTable = "flink_delete_src_test" + RandomUtils.randomString();
+    dstTable = "flink_delete_dst_test" + RandomUtils.randomString();
 
     ExecutorService executor = Executors.newSingleThreadExecutor();
 
@@ -79,14 +119,13 @@ public class TiKVDeduplicateTest extends FlinkTestBase {
     properties.put(STREAMING_CODEC.key(), STREAMING_CODEC_JSON);
     properties.put("tidb.streaming.kafka.bootstrap.servers", "localhost:9092");
     properties.put("tidb.streaming.kafka.topic", "tidb_test");
-    properties.put("tidb.streaming.kafka.group.id", "group_dep");
+    properties.put("tidb.streaming.kafka.group.id", kafkaGroup);
     properties.put(IGNORE_PARSE_ERRORS.key(), "true");
     // sink
     properties.put(SINK_IMPL.key(), TIKV.name());
-    properties.put(SINK_BUFFER_SIZE.key(), "2");
+    properties.put(SINK_BUFFER_SIZE.key(), "1");
     properties.put(WRITE_MODE.key(), TiDBWriteMode.UPSERT.name());
     properties.put(SINK_TRANSACTION.key(), SinkTransaction.MINIBATCH.name());
-    properties.put(DEDUPLICATE.key(), Boolean.toString(true));
     properties.put(DELETE_ENABLE.key(), Boolean.toString(true));
 
     FutureTask<Optional<JobClient>> task =
@@ -99,11 +138,11 @@ public class TiKVDeduplicateTest extends FlinkTestBase {
                   StreamTableEnvironment.create(env, settings);
               // init catalog and create dstTable
               TiDBCatalog tiDBCatalog =
-                  initTiDBCatalog(dstTable, TABLE_UK, tableEnvironment, properties);
+                  initTiDBCatalog(dstTable, flinkDeleteTable, tableEnvironment, properties);
               // create src table
               tiDBCatalog.sqlUpdate(
                   String.format("DROP TABLE IF EXISTS `%s`.`%s`", DATABASE_NAME, srcTable));
-              tiDBCatalog.sqlUpdate(String.format(TABLE_UK, DATABASE_NAME, srcTable));
+              tiDBCatalog.sqlUpdate(String.format(flinkDeleteTable, DATABASE_NAME, srcTable));
 
               TableResult tableResult =
                   tableEnvironment.executeSql(
@@ -118,25 +157,31 @@ public class TiKVDeduplicateTest extends FlinkTestBase {
     Thread.sleep(20000);
 
     JobClient jobClient = task.get().orElseThrow(IllegalStateException::new);
-    // insert (1,1)
-    testDatabase
-        .getClientSession()
-        .sqlUpdate(String.format("insert into `%s`.`%s` values('1','1')", DATABASE_NAME, srcTable));
-    // delete (1,1)
-    testDatabase
-        .getClientSession()
-        .sqlUpdate(String.format("delete from `%s`.`%s` where c1 = '1'", DATABASE_NAME, srcTable));
-    // trigger row flush
+
+    // insert 4 rows in src, flush 3 insert to dst
     testDatabase
         .getClientSession()
         .sqlUpdate(
             String.format(
-                "insert into `%s`.`%s` values('3','3'),('4','4')", DATABASE_NAME, srcTable));
+                "insert into `%s`.`%s` values('1','1'),('2','2'),('3','3'),('4','4')",
+                DATABASE_NAME, srcTable));
     Thread.sleep(20000);
-
-    // +I(3,3) will do 2pc
     Assert.assertEquals(
-        1, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
+        4, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, srcTable));
+    Assert.assertEquals(
+        3, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
+    // delete 3 rows in src , flush 1 insert and 2 delete to dst
+    testDatabase
+        .getClientSession()
+        .sqlUpdate(
+            String.format(
+                "delete from `%s`.`%s` where c1 = '1' or c1 = '2' or c1 = '3'",
+                DATABASE_NAME, srcTable));
+    Thread.sleep(20000);
+    Assert.assertEquals(
+        1, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, srcTable));
+    Assert.assertEquals(
+        result, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
 
     jobClient.cancel();
   }
