+++ flink/flink-1.14/src/test/java/io/tidb/bigdata/flink/tidb/sink/TiKVDeduplicateTest.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,16 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tidb.bigdata.flink.tidb.stream.delete;
 
+package io.tidb.bigdata.flink.tidb.sink;
+
+import static io.tidb.bigdata.flink.connector.TiDBOptions.DEDUPLICATE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.DELETE_ENABLE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.IGNORE_PARSE_ERRORS;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_BUFFER_SIZE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_IMPL;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SINK_TRANSACTION;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC;
-import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_CANAL_JSON;
-import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_CRAFT;
+import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_CODEC_JSON;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.STREAMING_SOURCE;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.SinkImpl.TIKV;
 import static io.tidb.bigdata.flink.connector.TiDBOptions.WRITE_MODE;
@@ -33,9 +34,8 @@ import io.tidb.bigdata.flink.connector.TiDBOptions.SinkTransaction;
 import io.tidb.bigdata.flink.tidb.FlinkTestBase;
 import io.tidb.bigdata.test.IntegrationTest;
 import io.tidb.bigdata.test.RandomUtils;
+import io.tidb.bigdata.test.StreamIntegrationTest;
 import io.tidb.bigdata.tidb.TiDBWriteMode;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ExecutorService;
@@ -50,71 +50,44 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized.Parameters;
 
-/** test for different codec, json has been test in {@link TiKVDeleteTest} */
-@Category(IntegrationTest.class)
-@RunWith(org.junit.runners.Parameterized.class)
-public class TiKVDeleteCodecTest extends FlinkTestBase {
+@Category(StreamIntegrationTest.class)
+public class TiKVDeduplicateTest extends FlinkTestBase {
+
   private String srcTable;
 
   private String dstTable;
 
-  public TiKVDeleteCodecTest(
-      String flinkDeleteTable, int result, String kafkaGroup, String topic, String streamingCodec) {
-    this.flinkDeleteTable = flinkDeleteTable;
-    this.result = result;
-    this.kafkaGroup = kafkaGroup;
-    this.topic = topic;
-    this.streamingCodec = streamingCodec;
-  }
-
-  @Parameters(
-      name =
-          "{index}: FlinkDeleteTable={0}, Result={1} ,KafkaGroup={2},Topic={3},StreamingCodec={4}")
-  public static Collection<Object[]> data() {
-    return Arrays.asList(
-        new Object[][] {
-          {TABLE_PK, 2, "group_c_1", "tidb_test_craft", STREAMING_CODEC_CRAFT},
-          {TABLE_PK, 2, "group_c_2", "tidb_test_canal_json", STREAMING_CODEC_CANAL_JSON},
-        });
-  }
-
-  private final String flinkDeleteTable;
-  private final int result;
-  private final String kafkaGroup;
-  private final String topic;
-  private final String streamingCodec;
-
-  private static final String TABLE_PK =
+  private static final String TABLE_UK =
       "CREATE TABLE IF NOT EXISTS `%s`.`%s`\n"
           + "(\n"
-          + "    c1  bigint(20),\n"
-          + "    c2  varchar(255),\n"
-          + "    PRIMARY KEY (`c1`) \n"
+          + "    c1  bigint(20) NULL DEFAULT NULL,\n"
+          + "    c2  bigint(20) NOT NULL,\n"
+          + "    UNIQUE INDEX `uniq_1`(`c1`) USING BTREE,\n"
+          + "    UNIQUE INDEX `uniq_2`(`c2`) USING BTREE\n"
           + ")";
 
   @Test
-  public void testDelete() throws Exception {
-    srcTable = "flink_delete_src_test" + RandomUtils.randomString();
-    dstTable = "flink_delete_dst_test" + RandomUtils.randomString();
+  public void testDeduplicate() throws Exception {
+    srcTable = "flink_deduplicate_src_test" + RandomUtils.randomString();
+    dstTable = "flink_deduplicate_dst_test" + RandomUtils.randomString();
 
     ExecutorService executor = Executors.newSingleThreadExecutor();
 
     Map<String, String> properties = defaultProperties();
     // source
     properties.put(STREAMING_SOURCE.key(), "kafka");
-    properties.put(STREAMING_CODEC.key(), streamingCodec);
+    properties.put(STREAMING_CODEC.key(), STREAMING_CODEC_JSON);
     properties.put("tidb.streaming.kafka.bootstrap.servers", "localhost:9092");
-    properties.put("tidb.streaming.kafka.topic", topic);
-    properties.put("tidb.streaming.kafka.group.id", kafkaGroup);
+    properties.put("tidb.streaming.kafka.topic", "tidb_test");
+    properties.put("tidb.streaming.kafka.group.id", "group_dep");
     properties.put(IGNORE_PARSE_ERRORS.key(), "true");
     // sink
     properties.put(SINK_IMPL.key(), TIKV.name());
-    properties.put(SINK_BUFFER_SIZE.key(), "1");
+    properties.put(SINK_BUFFER_SIZE.key(), "2");
     properties.put(WRITE_MODE.key(), TiDBWriteMode.UPSERT.name());
     properties.put(SINK_TRANSACTION.key(), SinkTransaction.MINIBATCH.name());
+    properties.put(DEDUPLICATE.key(), Boolean.toString(true));
     properties.put(DELETE_ENABLE.key(), Boolean.toString(true));
 
     FutureTask<Optional<JobClient>> task =
@@ -127,11 +100,11 @@ public class TiKVDeleteCodecTest extends FlinkTestBase {
                   StreamTableEnvironment.create(env, settings);
               // init catalog and create dstTable
               TiDBCatalog tiDBCatalog =
-                  initTiDBCatalog(dstTable, flinkDeleteTable, tableEnvironment, properties);
+                  initTiDBCatalog(dstTable, TABLE_UK, tableEnvironment, properties);
               // create src table
               tiDBCatalog.sqlUpdate(
                   String.format("DROP TABLE IF EXISTS `%s`.`%s`", DATABASE_NAME, srcTable));
-              tiDBCatalog.sqlUpdate(String.format(flinkDeleteTable, DATABASE_NAME, srcTable));
+              tiDBCatalog.sqlUpdate(String.format(TABLE_UK, DATABASE_NAME, srcTable));
 
               TableResult tableResult =
                   tableEnvironment.executeSql(
@@ -146,31 +119,25 @@ public class TiKVDeleteCodecTest extends FlinkTestBase {
     Thread.sleep(20000);
 
     JobClient jobClient = task.get().orElseThrow(IllegalStateException::new);
-    // insert 4 rows in src, flush 3 insert to dst
+    // insert (1,1)
+    testDatabase
+        .getClientSession()
+        .sqlUpdate(String.format("insert into `%s`.`%s` values('1','1')", DATABASE_NAME, srcTable));
+    // delete (1,1)
+    testDatabase
+        .getClientSession()
+        .sqlUpdate(String.format("delete from `%s`.`%s` where c1 = '1'", DATABASE_NAME, srcTable));
+    // trigger row flush
     testDatabase
         .getClientSession()
         .sqlUpdate(
             String.format(
-                "insert into `%s`.`%s` values('1','1'),('2','2'),('3','3'),('4','4')",
-                DATABASE_NAME, srcTable));
+                "insert into `%s`.`%s` values('3','3'),('4','4')", DATABASE_NAME, srcTable));
     Thread.sleep(20000);
-    Assert.assertEquals(
-        4, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, srcTable));
-    Assert.assertEquals(
-        3, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
 
-    // delete 3 rows in src , flush 1 insert and 2 delete to dst
-    testDatabase
-        .getClientSession()
-        .sqlUpdate(
-            String.format(
-                "delete from `%s`.`%s` where c1 = '1' or c1 = '2' or c1 = '3'",
-                DATABASE_NAME, srcTable));
-    Thread.sleep(20000);
+    // +I(3,3) will do 2pc
     Assert.assertEquals(
-        1, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, srcTable));
-    Assert.assertEquals(
-        result, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
+        1, testDatabase.getClientSession().queryTableCount(DATABASE_NAME, dstTable));
 
     jobClient.cancel();
   }
